On Mac skip killing PID
Intended to fix on Mac when using multitail to start multiple programs,
and handle properly when they exit. Example:
  multitail -r 1 -l "foobar" -r 1 -L "foobar -N"
Without this patch, it prints out:

 --*- multitail 6.4.3 (C) 2003-2014 by folkert@vanheusden.com -*--

The following problem occured:
-----------------------------
First Problem stopping child process with PID 39185 (SIGTERM).
If this is a bug, please report the following information:
The last system call returned: 1 which means "Operation not permitted"
 --*- multitail 6.4.3 (C) 2003-2014 by folkert@vanheusden.com -*--

The following problem occured:
-----------------------------
First Problem stopping child process with PID 39185 (SIGTERM).
If this is a bug, please report the following information:
The last system call returned: 1 which means "Operation not permitted"