Fix HTTP status codes colours in the "apache" scheme
In the Apache log file regexps the double quote is necessary to match
the correct number as the HTTP status code, but then cs_re applies the
colour to the entire matching string, including the double quote, which
is not the intended behaviour.  Use cs_re_s instead to fix this.